pyproject.toml
[project]
name = "borg2mqtt"
version = "0.1.1"
description = "Send borg backup updates over MQTT"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE.txt"}
keywords = ["borg", "mqtt", "hass"]  # Optional

authors = [
  {name = "Easton Potokar", email = "contagon6@gmail.com" } # Optional
]
maintainers = [
  {name = "Easton Potokar", email = "contagon6@gmail.com" } # Optional
]

classifiers = [  # Optional
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 4 - Beta",

  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
  "paho-mqtt", 
  "platformdirs", 
  "python-slugify",
  "pyyaml",
]

[project.optional-dependencies]
dev = ["check-manifest", "pre-commit"]
test = ["pytest"]

[project.urls]
"Homepage" = "https://github.com/contagon/borg2mqtt"
"Bug Reports" = "https://github.com/contagon/borg2mqtt/issues"
"Source" = "https://github.com/contagon/borg2mqtt"

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
borg2mqtt = "borg2mqtt:run_borg2mqtt"
